Settings and Edit pages added; fix Edit page; Add save for settings; file read still broken

+++ ViewModels/RecordsVM.cs
@@ -2,8 +2,10 @@
 using AverageAssistant.Models;
 using AverageAssistant.Services;
 using AverageAssistant.ViewModels;
+using AverageAssistant.Views;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using CommunityToolkit.Mvvm.Messaging;
 using JsonManagement;
 namespace AverageAssistant.RecordsVM;
 
@@ -16,21 +18,57 @@ public partial class RecordVM:ObservableObject
         Model = model;
         var Calculator = new AverageCalculator();
         Model.Average = ((IAverageSystem)Calculator).UsedAverageSystem(Model, Model, Model);
+
+
     }
 
     public async Task FileHandling()
     {
         var FileHandler = new JsonManager();
-        await ((IJsonManager)FileHandler).CreateFileForFlags(Model);
-        Model = await ((IJsonManager)FileHandler).ReadFromFileInput();
+
+        var inputRecord = await ((IJsonManager)FileHandler).ReadFromFile();
+
+        if (inputRecord != null)
+        {
+            Model.Grade = inputRecord.Grade;
+            Model.SubjectName = inputRecord.SubjectName;
+            Model.UsersGrades = inputRecord.UsersGrades;
+            Model.NumberOfLessons = inputRecord.NumberOfLessons;
+        }
+
+        OnPropertyChanged(nameof(Model));
+
     }
+
+
 
     [RelayCommand]
     public async Task DeleteFile()
     {
-        var deleteFile = new JsonManager();
+        bool answer = await Shell.Current.DisplayAlert("Warning!", "Are you sure that you want to delete this object?", "Delete", "Cancel");
 
-        await ((IJsonManager)deleteFile).DeleteFile(Model);
+        if (answer)
+        {
+            var deleteFile = new JsonManager();
+            await ((IJsonManager)deleteFile).DeleteFile(Model);
+        }
+    }
+
+    [RelayCommand]
+    public async Task GoToEdit()
+    {
+        var RecordForEdit = new Record()
+        {
+            Grade = Model.Grade,
+            SubjectName = Model.SubjectName,
+            UsersGrades = Model.UsersGrades,
+            NumberOfLessons = Model.NumberOfLessons,
+            SelectedAverageSystem = Model.SelectedAverageSystem
+        };
+
+        WeakReferenceMessenger.Default.Send(RecordForEdit);
+
+        await Shell.Current.GoToAsync(nameof(EditPage));
     }
 
     public string? AverageDisplay => Model.AverageDisplay;
@@ -46,8 +84,5 @@ public partial class RecordVM:ObservableObject
     public bool IsGradeVisible => Model.IsGradeVisible;
     public bool IsAverageVisible => Model.IsAverageVisible;
     public bool IsNrGradesWarningVisible => Model.IsNrGradesWarningVisible;
-
-    
-
 }
